VisPlotBench/requirements.txt
# Required dependencies
pandas>=2.2.2,<3.0.0
tqdm>=4.66.2,<5.0.0
nbformat>=5.10.4,<6.0.0
natsort>=8.4.0,<9.0.0
requests>=2.31.0,<3.0.0
tiktoken>=0.7.0,<0.8.0
omegaconf>=2.3.0,<3.0.0
python-dotenv>=1.0.1,<2.0.0
nbconvert>=7.16.3,<8.0.0
jupyter>=1.0.0,<2.0.0
matplotlib>=3.10.1,<4.0.0
jupyter-contrib-nbextensions>=0.7.0,<0.8.0
notebook==6.4.12
numpy>=1.26.4,<2.0.0
scipy>=1.13.0,<2.0.0
seaborn>=0.13.2,<0.14.0
plotly>=5.22.0,<6.0.0
kaleido==0.2.1
python-docx>=1.1.2,<2.0.0
datasets>=3.1.0,<4.0.0
fire>=0.7.0,<0.8.0
vllm
vl-convert-python==1.8.0
wurlitzer
geckodriver
cairosvg
pdf2image
playwright
webdriver-manager==4.0.2
chartify==5.0.1
bokeh==3.7.3
holoviews==1.21.0
pyvista==0.45.2
seaborn==0.13.2
pyecharts==2.0.8
altair==5.5.0
missingno==0.5.2
selenium==4.35.0
snapshot-selenium==0.0.2
tabulate